Websocket (#78)
* Save answers

* Websocket send current item

* Return PollItemDto instead of PollItem

* Save Answers

* Update Livepoll.postman_collection.json

* Code optimizations

* Fix typo

+++ src/main/kotlin/de/livepoll/api/entity/dto/PollDtoOut.kt
@@ -8,5 +8,5 @@ data class PollDtoOut(
         val startDate: Date,
         val endDate: Date,
         val slug: String,
-        var currentItem: Int?
+        var currentItem: Long?
 )
